feat: add router tests and moved common functions to root utils folder
