+++ test/__tests__/data/facebook_conversions.json
@@ -651,5 +651,103 @@
       },
       "files": {}
     }
+  },
+  {
+    "description": "Track event with standard event product list viewed without products array",
+    "input": {
+      "message": {
+        "anonymousId": "c82cbdff-e5be-4009-ac78-cdeea09ab4b1",
+        "channel": "web",
+        "context": {
+          "device": {
+            "id": "df16bffa-5c3d-4fbb-9bce-3bab098129a7R",
+            "manufacturer": "Xiaomi",
+            "model": "Redmi 6",
+            "name": "xiaomi"
+          },
+          "network": {
+            "carrier": "Banglalink"
+          },
+          "os": {
+            "name": "android",
+            "version": "8.1.0"
+          },
+          "screen": {
+            "height": "100",
+            "density": 50
+          },
+          "traits": {
+            "email": "    aBc@gmail.com   ",
+            "address": {
+              "zip": 1234
+            },
+            "anonymousId": "c82cbdff-e5be-4009-ac78-cdeea09ab4b1"
+          }
+        },
+        "event": "product list viewed",
+        "integrations": {
+          "All": true
+        },
+        "message_id": "a80f82be-9bdc-4a9f-b2a5-15621ee41df8",
+        "properties": {
+          "revenue": 400,
+          "additional_bet_index": 0,
+          "category": "randomCategory"
+        },
+        "timestamp": "2023-10-15T15:46:51.693229+05:30",
+        "type": "track"
+      },
+      "destination": {
+        "Config": {
+          "limitedDataUsage": true,
+          "blacklistPiiProperties": [
+            {
+              "blacklistPiiProperties": "",
+              "blacklistPiiHash": false
+            }
+          ],
+          "accessToken": "09876",
+          "datasetId": "dummyID",
+          "eventsToEvents": [
+            {
+              "from": "",
+              "to": ""
+            }
+          ],
+          "eventCustomProperties": [
+            {
+              "eventCustomProperties": ""
+            }
+          ],
+          "removeExternalId": true,
+          "whitelistPiiProperties": [
+            {
+              "whitelistPiiProperties": ""
+            }
+          ],
+          "actionSource": "website"
+        },
+        "Enabled": true
+      }
+    },
+    "output": {
+      "version": "1",
+      "type": "REST",
+      "method": "POST",
+      "endpoint": "https://graph.facebook.com/v17.0/dummyID/events?access_token=09876",
+      "headers": {},
+      "params": {},
+      "body": {
+        "JSON": {},
+        "XML": {},
+        "JSON_ARRAY": {},
+        "FORM": {
+          "data": [
+            "{\"user_data\":{\"em\":\"48ddb93f0b30c475423fe177832912c5bcdce3cc72872f8051627967ef278e08\",\"zp\":\"03ac674216f3e15c761ee1a5e255f067953623c8b388b4459e13f978d7c846f4\"},\"event_name\":\"ViewContent\",\"event_time\":1697365011,\"action_source\":\"website\",\"custom_data\":{\"revenue\":400,\"additional_bet_index\":0,\"category\":\"randomCategory\",\"content_ids\":[\"randomCategory\"],\"contents\":[{\"id\":\"randomCategory\",\"quantity\":1}],\"content_type\":\"product_group\",\"content_category\":\"randomCategory\",\"currency\":\"USD\",\"value\":400}}"
+          ]
+        }
+      },
+      "files": {}
+    }
   }
 ]
