+++ test/__tests__/data/facebook_conversions_router_input.json
@@ -1,0 +1,158 @@
+[
+  {
+    "metadata": {
+      "jobId": 1
+    },
+    "message": {
+      "anonymousId": "c82cbdff-e5be-4009-ac78-cdeea09ab4b1",
+      "channel": "web",
+      "context": {
+        "device": {
+          "id": "df16bffa-5c3d-4fbb-9bce-3bab098129a7R",
+          "manufacturer": "Xiaomi",
+          "model": "Redmi 6",
+          "name": "xiaomi"
+        },
+        "network": {
+          "carrier": "Banglalink"
+        },
+        "os": {
+          "name": "android",
+          "version": "8.1.0"
+        },
+        "screen": {
+          "height": "100",
+          "density": 50
+        },
+        "traits": {
+          "email": "    aBc@gmail.com   ",
+          "address": {
+            "zip": 1234
+          },
+          "anonymousId": "c82cbdff-e5be-4009-ac78-cdeea09ab4b1"
+        }
+      },
+      "event": "spin_result",
+      "integrations": {
+        "All": true
+      },
+      "message_id": "a80f82be-9bdc-4a9f-b2a5-15621ee41df8",
+      "properties": {
+        "revenue": 400,
+        "additional_bet_index": 0
+      },
+      "timestamp": "2023-10-15T15:46:51.693229+05:30",
+      "type": "track"
+    },
+    "destination": {
+      "Config": {
+        "limitedDataUsage": true,
+        "blacklistPiiProperties": [
+          {
+            "blacklistPiiProperties": "",
+            "blacklistPiiHash": false
+          }
+        ],
+        "accessToken": "09876",
+        "datasetId": "dummyID",
+        "eventsToEvents": [
+          {
+            "from": "",
+            "to": ""
+          }
+        ],
+        "eventCustomProperties": [
+          {
+            "eventCustomProperties": ""
+          }
+        ],
+        "removeExternalId": true,
+        "whitelistPiiProperties": [
+          {
+            "whitelistPiiProperties": ""
+          }
+        ],
+        "actionSource": "website"
+      },
+      "Enabled": true
+    }
+  },
+  {
+    "metadata": {
+      "jobId": 2
+    },
+    "message": {
+      "anonymousId": "c82cbdff-e5be-4009-ac78-cdeea09ab4b1",
+      "channel": "web",
+      "context": {
+        "device": {
+          "id": "df16bffa-5c3d-4fbb-9bce-3bab098129a7R",
+          "manufacturer": "Xiaomi",
+          "model": "Redmi 6",
+          "name": "xiaomi"
+        },
+        "network": {
+          "carrier": "Banglalink"
+        },
+        "os": {
+          "name": "android",
+          "version": "8.1.0"
+        },
+        "screen": {
+          "height": "100",
+          "density": 50
+        },
+        "traits": {
+          "email": "    aBc@gmail.com   ",
+          "address": {
+            "zip": 1234
+          },
+          "anonymousId": "c82cbdff-e5be-4009-ac78-cdeea09ab4b1"
+        }
+      },
+      "event": "products searched",
+      "integrations": {
+        "All": true
+      },
+      "message_id": "a80f82be-9bdc-4a9f-b2a5-15621ee41df8",
+      "properties": {
+        "revenue": 400,
+        "additional_bet_index": 0
+      },
+      "timestamp": "2023-10-15T15:46:51.693229+05:30",
+      "type": "track"
+    },
+    "destination": {
+      "Config": {
+        "limitedDataUsage": true,
+        "blacklistPiiProperties": [
+          {
+            "blacklistPiiProperties": "",
+            "blacklistPiiHash": false
+          }
+        ],
+        "accessToken": "09876",
+        "datasetId": "dummyID",
+        "eventsToEvents": [
+          {
+            "from": "",
+            "to": ""
+          }
+        ],
+        "eventCustomProperties": [
+          {
+            "eventCustomProperties": ""
+          }
+        ],
+        "removeExternalId": true,
+        "whitelistPiiProperties": [
+          {
+            "whitelistPiiProperties": ""
+          }
+        ],
+        "actionSource": "website"
+      },
+      "Enabled": true
+    }
+  }
+]
